debug
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/OrthoMCLWorkflow/trunk@68596 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ Main/lib/perl/WorkflowSteps/MapAndInsertPrevSeqIDs.pm
@@ -40,6 +40,8 @@ sub run {
     my $oldAbbrev = $abbrev;
     $oldAbbrev = $oldAbbrevs->{$releaseDir} if ($oldAbbrevs->{$releaseDir});
 
+    next unless -e  "$workflowDataDir/$inputDir/$releaseDir/$oldAbbrev.fasta.gz";
+
     unlink("tmpAbbrevMap") if -e "tmpAbbrevMap";
     open(F, ">tmpAbbrevMap");
     print F "$oldAbbrev $abbrev\n";
